pyproject.toml
[project]
name = "airflow-edge-executor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow>=3.1.0",
    "apache-airflow-providers-celery>=3.12.4",
    "apache-airflow-providers-edge3>=1.3.1",
    "pymysql>=1.1.2",
]
